requirements.txt
aiohttp==3.9.1
beautifulsoup4==4.12.2
cerebras-cloud-sdk>=1.5.0
flask==3.0.0
python-dotenv==1.0.0
pydantic==2.5.0
markdown==3.5.1
pygments==2.17.2
httpx>=0.24.0cd c